+++ packages/progressive-review/app/src/software-map/SoftwareMap.test.ts
@@ -3,9 +3,12 @@ import { readFileSync } from "node:fs";
 import type { Edge as ReactFlowEdge } from "@xyflow/react";
 import { describe, expect, it, vi } from "vitest";
 
-import { testReviewSession } from "../review-session-test-utils";
 import { collapseInlineC4Node, projectInlineC4 } from "./c4-projection";
 import { defineSoftwareModel } from "./model";
+import {
+  initialSoftwareMapExpandedNodeIds,
+  seedSoftwareMapDefaultExpandedNodeIds,
+} from "./software-map-navigation-state";
 import {
   C4LayoutQueue,
   C4_MAP_HOTKEY_GROUPS,
@@ -22,7 +25,6 @@ import {
   c4PreviousInlineLayoutForRelationships,
   c4SpatialDirectionForKey,
   c4ViewportForNodeReveal,
-  clearSoftwareMapNavigationStateForTests,
   createC4MapFlow,
   createC4MapFlowFromLayout,
   findSpatialC4Node,
@@ -30,22 +32,16 @@ import {
   fitC4MapView,
   focusC4MapNode,
   focusC4MapNodeAndKeyboard,
-  initialSoftwareMapExpandedNodeIds,
   parentSoftwareMapNodeId,
   positionC4EdgeLabels,
-  rememberSoftwareMapNavigationState,
-  restoreSoftwareMapNavigationState,
   runInlineC4Layout,
   runSerializedC4Layout,
   scheduleC4NodeMeasurements,
-  seedSoftwareMapDefaultExpandedNodeIds,
   selectedSoftwareMapNodeIdForNodes,
   shouldApplySoftwareMapModifiedOnly,
   shouldAutoFocusC4MapKeyboardTarget,
   shouldShowSoftwareMapFloatingActions,
-  softwareMapAncestorPaths,
   softwareMapChildNodeIdForDrill,
-  softwareMapNavigationKey,
   softwareMapNodeDiffPeeks,
   softwareMapNodeForKeyboardExpansion,
   softwareMapNodeIdForDrill,
@@ -261,18 +257,6 @@ describe("SoftwareMap inline C4 helpers", () => {
     });
     expect(expanded.codeElements).toEqual([]);
     expect(expanded.coverageClaims).toHaveLength(2);
-  });
-
-  it("derives ancestors for map-backed side peek focus requests", () => {
-    expect(
-      softwareMapAncestorPaths(
-        "progressiveReview.reviewApp.databaseLens.persistOperation",
-      ),
-    ).toEqual([
-      "progressiveReview",
-      "progressiveReview.reviewApp",
-      "progressiveReview.reviewApp.databaseLens",
-    ]);
   });
 
   it("keeps authored-only maps visible when modified-only debug filtering is enabled", () => {
@@ -1111,166 +1095,6 @@ describe("SoftwareMap inline C4 helpers", () => {
     expect(styles).toMatch(
       /\.software-map-canvas\s*{[^}]*min-height:\s*0;[^}]*overflow:\s*hidden;/s,
     );
-  });
-
-  it("persists selected node and expanded node ids by model identity", () => {
-    const session = testReviewSession();
-    clearSoftwareMapNavigationStateForTests(session);
-    const key = softwareMapNavigationKey({
-      title: "CI SoftwareMap",
-      view: "inline",
-    });
-
-    rememberSoftwareMapNavigationState(session, key, {
-      modelKey: "model:a",
-      expandedNodeIds: ["devFastCi", "devFastCi.ciWorker"],
-      selectedNodeId: "devFastCi.ciWorker",
-      expanded: true,
-    });
-
-    expect(restoreSoftwareMapNavigationState(session, key, "model:a")).toEqual({
-      modelKey: "model:a",
-      expandedNodeIds: ["devFastCi", "devFastCi.ciWorker"],
-      selectedNodeId: "devFastCi.ciWorker",
-      expanded: true,
-    });
-    expect(restoreSoftwareMapNavigationState(session, key, "model:b")).toEqual({
-      modelKey: "model:b",
-      expandedNodeIds: [],
-      selectedNodeId: null,
-      expanded: false,
-    });
-  });
-
-  it("defaults every non-component expandable node to expanded", () => {
-    const model = defineSoftwareModel({
-      systems: {
-        app: {
-          containers: {
-            web: {
-              components: {
-                ui: {
-                  codeElements: {
-                    render: {
-                      sourceRanges: [
-                        { file: "src/example.ts", fromLine: 1, toLine: 1 },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-          dataStores: {
-            graph: {
-              tables: {
-                nodes: { schema: { id: { type: "string", pk: true } } },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    expect([...initialSoftwareMapExpandedNodeIds(model)].sort()).toEqual([
-      "app",
-      "app.graph",
-      "app.web",
-    ]);
-  });
-
-  it("seeds nested default expansion once the complete model is available", () => {
-    const initialModel = defineSoftwareModel({
-      systems: {
-        app: {
-          containers: {
-            web: {},
-          },
-        },
-      },
-    });
-    const completeModel = defineSoftwareModel({
-      systems: {
-        app: {
-          containers: {
-            web: {
-              components: {
-                ui: {
-                  codeElements: {
-                    render: {
-                      sourceRanges: [
-                        { file: "src/example.ts", fromLine: 1, toLine: 1 },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    const initialExpandedNodeIds =
-      initialSoftwareMapExpandedNodeIds(initialModel);
-    expect([...initialExpandedNodeIds]).toEqual(["app"]);
-
-    const expandedNodeIds = seedSoftwareMapDefaultExpandedNodeIds({
-      expandedNodeIds: initialExpandedNodeIds,
-      model: completeModel,
-      defaultExpansionActive: true,
-    });
-    expect([...expandedNodeIds].sort()).toEqual(["app", "app.web"]);
-
-    const projection = projectInlineC4({
-      model: completeModel,
-      expandedNodeIds,
-    });
-    expect(projection.nodes).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ id: "app", isExpanded: true }),
-        expect.objectContaining({ id: "app.web", isExpanded: true }),
-        expect.objectContaining({
-          id: "app.web.ui",
-          isExpanded: false,
-        }),
-      ]),
-    );
-    expect(projection.visibleNodeIds.has("app.web.ui.render")).toBe(false);
-  });
-
-  it("does not re-expand a default node after the user collapses it", () => {
-    const model = defineSoftwareModel({
-      systems: {
-        app: {
-          containers: {
-            web: {
-              components: {
-                ui: {
-                  codeElements: {
-                    render: {
-                      sourceRanges: [
-                        { file: "src/example.ts", fromLine: 1, toLine: 1 },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    expect(
-      [
-        ...seedSoftwareMapDefaultExpandedNodeIds({
-          expandedNodeIds: new Set(["app"]),
-          model,
-          defaultExpansionActive: false,
-        }),
-      ].sort(),
-    ).toEqual(["app"]);
   });
 
   it("updates the first map layout when resolved children and edges arrive", async () => {
