+++ packages/progressive-review/app/src/software-map/SoftwareMap.tsx
@@ -64,11 +64,6 @@ import {
   useReviewActions,
 } from "../review-context";
 import { useReviewInitialData } from "../review-initial-data-context";
-import {
-  forgetReviewUiState,
-  readReviewUiState,
-  writeReviewUiState,
-} from "../review-ui-state";
 import { useRightPanelResize } from "../side-panel-resizer";
 import { buildGraphTarget, targetKey } from "../target-fingerprint";
 import { useRegisterLiveDiagram } from "../thread-target-model";
@@ -78,7 +73,6 @@ import {
   type C4Projection,
   type ProjectedC4Relationship,
   collapseInlineC4Node,
-  isInlineC4Expandable,
   projectInlineC4,
 } from "./c4-projection";
 import {
@@ -94,6 +88,15 @@ import {
   SoftwareMapUnavailable,
   softwareMapCssLength,
 } from "./software-map-absence";
+import {
+  hasStoredSoftwareMapNavigationState,
+  initialSoftwareMapExpandedNodeIds,
+  rememberSoftwareMapNavigationState,
+  restoreSoftwareMapNavigationState,
+  seedSoftwareMapDefaultExpandedNodeIds,
+  softwareMapAncestorPaths,
+  softwareMapNavigationKey,
+} from "./software-map-navigation-state";
 import { refreshSoftwareMapArtifacts } from "./software-map-patch-client";
 import {
   relationshipEndpointsKey,
@@ -540,13 +543,6 @@ export interface C4SpatialNodePosition {
   height?: number;
 }
 
-interface SoftwareMapNavigationState {
-  modelKey: string | undefined;
-  expandedNodeIds: string[];
-  selectedNodeId: string | null;
-  expanded: boolean;
-}
-
 const ELEMENT_TYPE_LABELS: Record<SoftwareMapElementType, string> = {
   person: "Person",
   softwareSystem: "System",
@@ -671,10 +667,6 @@ export function runSerializedC4Layout<Result>(
   return c4LayoutQueue.run(task);
 }
 
-const softwareMapNavigationStateByKey = new Map<
-  string,
-  SoftwareMapNavigationState
->();
 function createSoftwareMapSignature() {
   let hash = 0x811c9dc5;
   const addText = (text: string) => {
@@ -813,159 +805,6 @@ export function shouldApplySoftwareMapModifiedOnly({
       softwareMapResolvedDataInputHasWork(resolvedDataInput),
     )
   );
-}
-
-export function softwareMapNavigationKey({
-  title,
-  view,
-  placeholderLabel = "Software map",
-}: {
-  title?: string;
-  view?: string;
-  placeholderLabel?: string;
-}) {
-  return [title ?? "", view ?? "", placeholderLabel].join("\u001f");
-}
-
-export function softwareMapAncestorPaths(path: string): string[] {
-  const parts = path.split(".");
-  const ancestors: string[] = [];
-  for (let index = 1; index < parts.length; index += 1) {
-    ancestors.push(parts.slice(0, index).join("."));
-  }
-  return ancestors;
-}
-
-function defaultSoftwareMapNavigationState(
-  modelKey: string | undefined,
-): SoftwareMapNavigationState {
-  return {
-    modelKey,
-    expandedNodeIds: [],
-    selectedNodeId: null,
-    expanded: false,
-  };
-}
-
-function cachedSoftwareMapNavigationState(session: ReviewSession, key: string) {
-  return (
-    softwareMapNavigationStateByKey.get(
-      softwareMapNavigationStorageKey(session, key),
-    ) ?? readStoredSoftwareMapNavigationState(session, key)
-  );
-}
-
-export function hasStoredSoftwareMapNavigationState(
-  session: ReviewSession,
-  key: string,
-  modelKey: string | undefined,
-) {
-  return cachedSoftwareMapNavigationState(session, key)?.modelKey === modelKey;
-}
-
-export function restoreSoftwareMapNavigationState(
-  session: ReviewSession,
-  key: string,
-  modelKey: string | undefined,
-): SoftwareMapNavigationState {
-  const cached = cachedSoftwareMapNavigationState(session, key);
-  if (!cached || cached.modelKey !== modelKey) {
-    return defaultSoftwareMapNavigationState(modelKey);
-  }
-  return {
-    modelKey,
-    expandedNodeIds: [...cached.expandedNodeIds],
-    selectedNodeId: cached.selectedNodeId,
-    expanded: cached.expanded,
-  };
-}
-
-export function initialSoftwareMapExpandedNodeIds(
-  model: NormalizedSoftwareModel | null | undefined,
-): Set<string> {
-  return new Set(
-    model?.elements
-      .filter(
-        (element) =>
-          element.type !== "component" && isInlineC4Expandable(element),
-      )
-      .map((element) => element.path) ?? [],
-  );
-}
-
-export function seedSoftwareMapDefaultExpandedNodeIds(input: {
-  expandedNodeIds: ReadonlySet<string>;
-  model: NormalizedSoftwareModel | null | undefined;
-  defaultExpansionActive: boolean;
-}): Set<string> {
-  if (!input.defaultExpansionActive) {
-    return new Set(input.expandedNodeIds);
-  }
-  const expandedNodeIds = new Set(input.expandedNodeIds);
-  for (const path of initialSoftwareMapExpandedNodeIds(input.model)) {
-    expandedNodeIds.add(path);
-  }
-  return expandedNodeIds;
-}
-
-export function rememberSoftwareMapNavigationState(
-  session: ReviewSession,
-  key: string,
-  state: SoftwareMapNavigationState,
-) {
-  softwareMapNavigationStateByKey.set(
-    softwareMapNavigationStorageKey(session, key),
-    {
-      ...state,
-      expandedNodeIds: [...state.expandedNodeIds],
-    },
-  );
-  writeStoredSoftwareMapNavigationState(session, key, state);
-}
-
-export function clearSoftwareMapNavigationStateForTests(
-  session: ReviewSession,
-) {
-  softwareMapNavigationStateByKey.clear();
-  if (typeof window !== "undefined") {
-    forgetReviewUiState("window", (key) =>
-      key.startsWith(session.storageKey("software-map-navigation")),
-    );
-  }
-}
-
-function softwareMapNavigationStorageKey(session: ReviewSession, key: string) {
-  return session.storageKey("software-map-navigation", key);
-}
-
-function readStoredSoftwareMapNavigationState(
-  session: ReviewSession,
-  key: string,
-): SoftwareMapNavigationState | null {
-  const parsed = readReviewUiState<JsonValue>(
-    "window",
-    softwareMapNavigationStorageKey(session, key),
-  );
-  if (!isJsonObject(parsed)) return null;
-  return {
-    modelKey: jsonString(jsonProperty(parsed, "modelKey")),
-    expandedNodeIds: (jsonArray(jsonProperty(parsed, "expandedNodeIds")) ?? [])
-      .map(jsonString)
-      .filter((entry): entry is string => entry !== undefined),
-    selectedNodeId: jsonString(jsonProperty(parsed, "selectedNodeId")) ?? null,
-    expanded: jsonProperty(parsed, "expanded") === true,
-  };
-}
-
-function writeStoredSoftwareMapNavigationState(
-  session: ReviewSession,
-  key: string,
-  state: SoftwareMapNavigationState,
-) {
-  writeReviewUiState("window", softwareMapNavigationStorageKey(session, key), {
-    ...state,
-    expandedNodeIds: [...state.expandedNodeIds],
-  });
 }
 
 export function findSpatialC4Node(
